feat: create and implement login and register, and more
A welcome screen is added. As well as auxiliary components user-init and anon-init.

Changes are made to the backend to facilitate the functionalities.

The login and register forms should be tested further. They seem to work properly, but I have observed some unexpected behaviours. They might be fixed already, but they might not.