dump: read desc
Refactored and restructured activity.component.ts, as well as its html.
  - Added chart settings (change x axis, show goal and limits)
  - Added selector for chart type (but does not work)
  - set the metric selector to real, selectable metrics.
  - the distribution is dynamically calculated

Created activity-streams.model.ts It is an object responsible for holding all the streams.

Greatly thinned down activity.model.ts, methods were moved to activity.service.ts

Created data-streams.ts, currently not in use. Its role would be to make it so that the streams in activity-streams can hold heterogeneous items, instead of all being forced to be strings

Created graph.service.ts, it is responsible for calvulating statistical metrics, as well as any operations that need to be made with the data for it to be able to be used in the graphs (at least in theory).