feat: implement DeletionService to handle delete
Whenever an Entity is deleted it should be through DeletionService. The delete method for each EntityService has been made package-private to prevent its use.

Deletes must go through DeletionService so that relationships between entities are properly handled.

In the limited testing done it appears to work properly, but more testing is needed for me to be certain it works as intended. Specifically, automated tests would be great.