fix: pace calculation and avgs
distance was sometimes 0 leading to an infinite pace. Now if pace==Long.MAX_VALUE, it takes the value of the previous or next value.

Avg pace was not working out properly, because "ratio of averages is not the same as the average of ratios", effectively, we should be doing a weighted average to consider different distance covered in each second. Since we are doing a normal average, it does not work properly.