feat: create GridItemService and mural dashboard
GridItemService will also be used for the user dashboard, which will more or less be the same as the mural dashboard.

GridItemService holds the current User/Mural that the dashboard is about. It also holds a list of members and routes, to have all the necessary data for Grid components centralized.

Currently there is only one Grid Component, though it's customizable. It's GridTableComponent. It shows a table. Currently all the logic on how to populate the tables is in the GridItemService. Moving it to the GridTableComponent should be considered, to clear up the Service.

The MuralDashboardComponent, when created, tells the service to load the necessary data. It also holds the GridItemComponents. These each query the service for the data they need. When the MuralDashboardComponent is destroyed, it signals the service to forget the loaded data.

More GridItemComponents should be added, and the Service should be refacored and reordered. The job's far from done.