chore: add mural home icon, highlight selected icon
This introduces a bug.

when going from /murals to /murals/owned, the icon changes from a house to an arrow, but its active status is not revoked. This happens because angular navigation is bypassed to not destroy the MuralComponent and allow data reuse.

To fix it, the data should be moved to a service, and angular navigation should be allowed to work normally.