feat: add activity visibility
There are 4 types:
- public: can be seen by anyone
- mural_public: can be seen by any mural the user is in, but not directly by other users
- mural_specific: can only be seen by a list of murals
- private: only the user can see it

There are two classes to implement this, the enum VisibilityType, and the class Visibility. The class is needed to hold the list of whitelisted murals.

Made sure the activity editing buttons only appear when the user owns the activity.

Deleted the possibility of accessing a list of specified activities

Updated GridItemService.fetchActivities to take a User/Mural, then return the activities of that User/Mural that are visible.

Deleted the line in the database initializer that set the mural's activities.