fix: repeated requests on activity load
When loading an activity, both the activity component and the study component call the service's get method to access the activity being loaded. This lead to two simultaneous requests for the same data. This was fixed by saving the activity in the service as a cache, and returning that if available and if it matches the requested activity.