fix: immediate logout on 401
Sometimes the backend crashes when decoding the JWT. This is interpreted as a failed authentication, and so 401 is returned. When a 401 was returned, the front immediately logged the user out. This is a hassle, as you're constantly being wrongfully logged out.

This has been solved by only logging the user out once they dismiss the alert informing them they've been logged out, and the alert's content has been modified to offer the option of reloading the page.