fix: updateVisibility not reflecting the change in front
Now when updating the visibility, backend returns the activity and we set it.

Some commented code gives the option to just change it locally, ignoring the backend response. With the original backend where it returns a boolean, this is better. I thought it's a good opportunity to update the cache.