chore: handle multiple alerts/confirms
Multiple alerts are allowed. If a copy of an open alertis shown, the open one is edited to have a multiplier.

Multiple confirms are NOT allowed. Temprorarily, and kinda for fun, they are tolerated if they are an exact copy. This is because of the issues with backend failing auth every once in a while. Once that is solved, trying to open a confirm while one is open should fail outright.