fix/ref/chore: bug fixes, code improvements...
There may be more minor changes than are listed here.

update logout to only log the user out in the frontend if the request to log them out in the backend was successful.

change alert when not logged in to a confirm offering to reload the page or log out (authInterceptor)

Add toasts on long loading screens suggesting sth might be wrong, and to reload

(backend) improve isLoggedIn and move it to LoginController

(frontend) improve isLoggedIn, using isFrontAuthenticated and isBackAuthenticated.

Change how CanActivate (authGuardService) checks the credentials (it's now async) due to the above changes.

change/update some comments

fix an issue where loading screen would not drop on activity not found. This modified ActivityService's cache slightly, and coincidentally solved the issue of the activity blacking out on cache expiration (though it may not have happened on this endpoint). A similar approach to this should be used for the mural cache.

add confirmation request on delete activity

fix layout not updating on change/delete/create route (mainly activityController)

center a bunch of modals

fix issue where deleting an activity from a route by clicking the trash icon in the route component gave 500 (securityConfig)

remove "not logged in" alert from profile component

delete {observe: 'response'} from userService.delete() as it's actually not needed. (The code is only accessed in the error, which already receives the status without observe:response)

fix error handling in activityService, and add better logging.

fix error handling in profileComponent

Delete stale code responsible for creating a mural with a list of users. This was dropped in favor of creating an empty mural and letting users join later. Some code remained. It has been deleted.

Hitting cancel on createMural did nothing.

Update labels for current/new password when changing it.