feat: implement join mural, changes in MuralsList
Fixed a potential issue where, due to the cache, the Owned and Member lists were not updating when a mural was created.

Cleaned MuralsListComponent by creating helper loadCategory methods to be called when we know changes have been made to a specific category. loadCategory updates the cache for that category in the service.

Other minor changes.