feat: add route visibility
Still need to apply visibility effects to some controller methods. That should be the next commit. This one just provides the basic functionality for fetching groups of routes. Fetching single routes and other CRUD operations are yet to be implemented

Moved linethrough-on-hover to global css

Created UtilsService, and moved changeVisibility functionality from activityController to it, to be easier to reuse.

Minor changes to User.java

Created ChangeVisibilityInterface, implemented by activity and route services (and any service that handles entities with visibility). It is needed as it allows passing the Service to changeVisibilityService, to be used without causing circular dependencies

Properly initialized the routes in databseInitializer