chore: add visibility concerns to other route/activity endpoints
Fixed a couple minor bugs. For one, when viewing an activity through a mural, the id was being fetched in an unclear way, and no record of which mural we're in was being made. This was fixed by changing the name of each id in the app.routes.ts. There are other small changes like this.