chore: FormattingService minor changes
Add method for distance formatting to FormattingService, and used it wherever distance in kms was being shown

Deleted 2 toHoursMinsSecs instances that were not being used and the one that was being used was moved to FormattingService

Consider in the future adding a method to format elevation. It may be redundant for now, but could be useful in the future.