refactor: delete some unneded relationships
Deleted the Many side of ManyToOne relationships, using database queries to fetch them. This makes logic simpler.

Also deleted some ManyToMany relationships that weren't actually needed.

More specifically:
- Mural->Activity: deleted since this relationship is managed via visibility.
- Activity->Mural:  deleted since Activity should not care what murals it's in

- Mural->Route: deleted since this relationship is managed via visibility.
- Route->Mural:  never existed outside visibility

-User--owns-->Mural: deleted since it can be fetched with muralRepo.findByOwner

-User->Activity: deleted since it can be fetched with activityRepo.findByUser

Deleting this las one seems to have caused issues in the databaseInitializer due to lazy loading of the activity's data points. I do not understand why that only became a problem after deleting the connection, but whatever. It was fixed by changing the init method from PostConstruct, which doesn't allow the Transactional tag, to an event listener that triggers when the app is ready. This allows use of Transactional, which keeps the session open, allowing access to the lazyly loaded fields. Maybe a better solution would be to just make the datapoints collection eagerly loaded. It should be considered in the future.