chore: minor changes and fixes (DB fix)
add a method to empty the database. For some fucking reason it was not being dropped as it should when the app is closed/opened

when a user is deleted non-private routes turn public

when a user is deleted an inheritor is selected automatically

when a user leaves a mural it is removed from activities that had it in their mural_specific visibility

forbid non-members from accessing any /murals/{id} children routes (/murals/{id}/activities, /routes, etc)

prevent non-members from fetching a mural

prevent mural visibility from being set to mural_public or mural_specific

ensure only Public murals appear on discover

Modified activity creation to not create activities with no datapoints. Additionally, add a check to activitySummary to throw an exception if the activity has no datapoints