chore: minor changes/fixes
fix an issue where mural codes where not being generated
fix an issue where the banner image was not being properly loaded. This and the last issue where introduced with the changes to the DatabaseInitializer (creation of beegInit)

fix issue where routes grid item didn't properly format distance. Also added options for time formatting, between : separators, h,m,s or h ' ''

fix an issue in new mural creation where the small for the aspect ratio was disappearing. It was caused by the errors object set manually being overwritten when the automatic error detection ran on blur.  To fix it, we made it into a validator. We decided to keep the check in onFileSelected, to have it up earlier and because it's useful for the notification