fix: private activities showing on routes
And other work on routes and their relationship with activities.

Deleted Route->Activity, substituting it with custom queries, since Activity already knows its Route.

Refactor route.owner to createdBy

Breaking change, endpoints that modify a route (delete, remove activity, add activity, etc) used to return the activity or activity list. This has been removed. They now return 204 no content. User is expected to do additional queries if required.
This was done because the visibility of the route's activities is messy to handle, so it's easier to just make the operation and leave the handling of the visibility to the actual fetching methods.
The fetching happens in two parts if you want the whole activity. You need to attack /routes/id for the route data, and /route/id/activities for its activities. This is meant to simplify logic and data transferred unnecessarily.

minor changes/fixes/refactor to Activity and Route repository custom queries

other minor changes/fixes