feat: add comparison page
It allows users to compare any number of activities with various tools.

Also added an option within routes to compare the activities in the route

Modified grid-item-service so that it returns the activity name and id when calculating records. Additionally, the records grid-item now shows the activity name and redirects to the activity on click

Had to modify activity-select to properly redirect the user to the comparison component, no matter how many activities they selected.

ActivityStudyComponent, refactored type of some variables to be more readable, and minor changes to the html text

Removed extra items in murals-dashboard component

Added a method/endpoint to fetch specific activities giving an id list. For visibility, the auth user is used, and/or a specified mural