refactor: move logic to service and more
Renamed UtilsService to AtraUtils and vastly changed its contents.

Created custom exceptions to be used instead of HttpException

Made ActivitySummary an entity that uses the same id as the Activity it summarizes. Whenever an activity is created through the service, an ActivitySummary is created to match it. When the activity is deleted, JPA deletes it through cascade and orphan removal

created isAdmin to substitute hasRole('ADMIN')

deleted DeletionService as it was no longer needed. It existed to avoid circular dependencies, but with direct repo access it's no longer needed.

Deleted some old unused code along with unused imports.